refactor : 타입 오류 해결
